requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
black==22.3.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.3.2
Deprecated==1.2.13
Django==4.0.3
djangorestframework==3.13.1
flake8==4.0.1
flower==1.0.0
gevent==21.12.0
greenlet==1.1.2
humanize==4.0.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
kombu==5.2.4
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.8
pytest==7.1.1
pytest-cov==3.0.0
pytest-django==4.5.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.2.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tornado==6.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
zope.event==4.5.0
zope.interface==5.4.0
